% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.chrTable.R
\name{chrtab}
\alias{chrtab}
\title{Create Baseline Characteristics Table}
\usage{
chrtab(
  variable,
  type,
  label,
  data,
  digits = 2,
  summary = "median",
  useNA = FALSE,
  na = "N/A",
  col1st = ""
)
}
\arguments{
\item{variable}{a vector of variables}

\item{type}{a vector of types either categorical or numerical (\code{"cat"} or \code{"num"})}

\item{label}{a vector of labels shown in the table}

\item{data}{data frame dataset}

\item{digits}{number of decimal points for numerical variable}

\item{summary}{three options are available for the numerical variable to be displayed; If \code{median}, median with min and max is calculated. If \code{median.iq}, median with 25 and 75 percentiles is calculated. If \code{mean}, mean with its standard deviation is calculated. Default is \code{"median"}.}

\item{useNA}{option whether or not \code{NULL} value is displayed for the categorical variable in the table. If \code{TRUE}, character in argument \code{na} is displayed for \code{NULL} value with its percentage. Default is \code{FALSE}.}

\item{na}{character to be displayed for \code{NULL} value if \code{useNA} is \code{TRUE}. Default is \code{"N/A"}.}

\item{col1st}{character label displayed in the 1st cell in the table}
}
\value{
summary matrix with calculated baseline characteristics
}
\description{
The function takes data and creates a table of baseline characteristics for the whole cohort.
}
\examples{
# example data is `mtcars` from the `datasets` package.  
chrtab(variable=c("mpg","gear"),type=c("num","cat"),
       label=c("MPG", "Gear"),data=mtcars,
       summary="median",useNA=FALSE,na="N/A",col1st="")

}
