% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{getMean}
\alias{getMean}
\title{Mean calculation with Standard Deviation}
\usage{
getMean(var, digits = 2)
}
\arguments{
\item{var}{is a vector of continuous variable}

\item{digits}{defines a decimal point}
}
\value{
A character vector with sd
}
\description{
Calculate mean and sd for continuous variables
}
\keyword{internal}
