% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{getMedian.mid}
\alias{getMedian.mid}
\title{Median Calculation with Q25 and Q75}
\usage{
getMedian.mid(var, digits = 2)
}
\arguments{
\item{var}{is a vector of continuous variable}

\item{digits}{defines a decimal point}
}
\value{
A character vector with median, q25, and q75
}
\description{
Calculate median, q25, and q75 for continuous variables
}
\keyword{internal}
