% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{getMedian}
\alias{getMedian}
\title{Median Calculation with Min and Max}
\usage{
getMedian(var, digits = 2)
}
\arguments{
\item{var}{is a vector of continuous variable}

\item{digits}{defines a decimal point}
}
\value{
A character vector with median, min, and max
}
\description{
Calculate median, min and max for continuous variables
}
\keyword{internal}
