% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{getNP}
\alias{getNP}
\title{Percentages Calculation with Frequencies}
\usage{
getNP(var, category, digits = 2)
}
\arguments{
\item{var}{is a categorical variable}

\item{category}{is a list of levels in categorical variable}

\item{digits}{defines a decimal point}
}
\value{
A character vector with sd
}
\description{
Calculate percentages for categorical variables with frequencies
}
\keyword{internal}
