% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.Cox.R
\name{mul.cox.reg}
\alias{mul.cox.reg}
\title{Create Multivariable Cox PH Regression Table}
\usage{
mul.cox.reg(time, status, covariate, label, type, reflevel = NULL, data)
}
\arguments{
\item{time}{a variable for time to event}

\item{status}{a variable for event (or censoring) indicator (1 indicates event, 0 is censored)}

\item{covariate}{a vector of variables included in the multivariable analysis}

\item{label}{a vector of labels shown in the table}

\item{type}{a vector of types either categorical or numerical (\code{"cat"} or \code{"num"})}

\item{reflevel}{a vector of reference levels for categorical variables. NA is used for numerical variable.}

\item{data}{data frame dataset}
}
\value{
matrix with estimates of coefficient, HR, and P-value
}
\description{
The function creates a result table for multivariable Cox PH regression. \code{veteran} dataset in the \code{survival} package is employed as an example data.
}
\examples{
library(survival)
data(veteran)
df <- veteran
mul.cox.reg(time="time", status="status", 
            covariate=c("age", "trt","celltype"), 
            label=c("Age", "Treatment", "Celltype"), 
            type=c("num","cat","cat"), 
            reflevel=c(NA, "1","squamous"), 
            data=df)
                            
}
