% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.Logistic.R
\name{mul.logistic.reg}
\alias{mul.logistic.reg}
\title{Create Multivariable Logistic Regression Table}
\usage{
mul.logistic.reg(
  outcome,
  covariate,
  label,
  type,
  reflevel = NULL,
  data,
  Odds.ratio = TRUE
)
}
\arguments{
\item{outcome}{a variable for event indicator (1 indicates event, 0 is censored)}

\item{covariate}{a vector of varaibles included in the univariable logistic analysis}

\item{label}{a vector of labels shown in the table for covariates}

\item{type}{a vector of types either categorical or continuous ("\code{cat}" or "\code{num}")}

\item{reflevel}{a vector of reference levels for categorical variables. \code{NA} is used for continuous variable.}

\item{data}{data frame dataset}

\item{Odds.ratio}{logical; if \code{TRUE}, odds ratio result is included. If \code{FALSE}, log odds ratio result is included and default is \code{TRUE}.}
}
\value{
matrix with estimates of coefficient, OR, and P-value
}
\description{
The function creates a result table of multivariable logistic regression.
}
\examples{
# Use mtcars dataset for example
df <- mtcars
df$mpg.cat <- ifelse(df$mpg > 18, 1, 0)

# Create variable, label, type, and reference information
r1 <- c("hp", "Horse power", "num", NA)
r2 <- c("wt", "Weight", "num", NA)
r3 <- c("gear", "Gear", "cat", "3")
input  <- rb(3)

mul.logistic.reg(outcome="mpg.cat", 
                 covariate=input[,"variable"],
                 label=input[,"label"],
                 type=input[,"type"],
                 reflevel=input[,"ref"], 
                 data=df, 
                 Odds.ratio=TRUE
                 )                  
}
