% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{na.handle.2}
\alias{na.handle.2}
\title{Transform "" levels as NA in factor}
\usage{
na.handle.2(data, variable, type, useNA, na)
}
\arguments{
\item{data}{is a dataset}

\item{variable}{is a variable}

\item{type}{is a type of that variable}

\item{useNA}{is whether or not to include NULL}

\item{na}{is how to express NULL values}
}
\value{
A updated vector with either NULL value or whatever values for NULL values such as "Unknown"
}
\description{
Calculate percentages for categorical variables with frequencies
}
\keyword{internal}
