% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{rb}
\alias{rb}
\title{Rowbind Vectors of Input}
\usage{
rb(x)
}
\arguments{
\item{x}{Number of vectors to be combined}
}
\value{
A matrix with following columns c(variable, label, type, reference) or c(variable, label, type, reference, order)
}
\description{
The function combines vectors of input by row-wise, where vector is composed of variable, label, type, and reference level.
}
\examples{
r1 <- c("mpg","MPG","num",NA,1)
r2 <- c("gear","Gear","cat","3",2)
rb(2)
}
