% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.Linear.R
\name{uni.linear.reg}
\alias{uni.linear.reg}
\title{Create Univariable Linear Regression Table}
\usage{
uni.linear.reg(outcome, covariate, label, type, reflevel = NULL, data)
}
\arguments{
\item{outcome}{a continous outcome variable}

\item{covariate}{a vector of variables included in the univariable linear analysis}

\item{label}{a vector of labels shown in the table for covariates}

\item{type}{a vector of types either categorical or numerical ("\code{cat}" or "\code{num}")}

\item{reflevel}{a vector of reference levels for categorical variables. \code{NA} is used for continuous variable.}

\item{data}{data frame dataset}
}
\value{
matrix with estimates of coefficients and P-value
}
\description{
The function creates a result table of univariable linear regression.
}
\examples{
# Use mtcars dataset for example
df <- mtcars
df$mpg.cat <- ifelse(df$mpg > 18, 1, 0)

# Create variable, label, type, and reference information
r1 <- c("hp", "Horse power", "num", NA)
r2 <- c("wt", "Weight", "num", NA)
r3 <- c("gear", "Gear", "cat", "3")
input  <- rb(3)

uni.linear.reg(outcome="mpg", 
               covariate=input[,"variable"],
               label=input[,"label"],
               type=input[,"type"],
               reflevel=input[,"ref"], 
               data=df
               )                
}
